Moved checkExists to @index/helpers.
